requirements.txt
# click, dnspython: No version, for Python2/3 compatibility
chacha20poly1305	>=0.0.3
click
configparser		>=4.0	; python_version < "3.0"
dnspython
enum34			>=1.1	; python_version < "3.4"
future			>=1
pathlib2		>=2.3	; python_version < "3.4"
pytz
requests		>=2.20,		<3
requests-toolbelt	>=1	; python_version > "2"
tzlocal
web.py			>=0.51
wsgilog			>=0.3
